+++ src/transactions/transferFeedUtils.ts
@@ -32,7 +32,6 @@ import {
 } from 'src/recipients/reducer'
 import { useSelector } from 'src/redux/hooks'
 import { useTokenInfoByAddress } from 'src/tokens/hooks'
-import { FeedTokenTransfer } from 'src/transactions/feed/TransferFeedItem'
 import {
   inviteTransactionsSelector,
   recentTxRecipientsCacheSelector,
@@ -43,6 +42,7 @@ import {
   TokenTransfer,
   TransactionStatus,
 } from 'src/transactions/types'
+import { isJumpstartTransaction } from 'src/transactions/utils'
 import Logger from 'src/utils/Logger'
 import { dataEncryptionKeySelector } from 'src/web3/selectors'
 
@@ -113,7 +113,7 @@ export function useTransactionRecipient(transfer: TokenTransfer): Recipient {
 }
 
 // Note: This hook is tested from src/transactions/feed/TransferFeedItem.test.ts
-export function useTransferFeedDetails(transfer: FeedTokenTransfer) {
+export function useTransferFeedDetails(transfer: TokenTransfer) {
   const { t } = useTranslation()
   const addressToDisplayName = useSelector(addressToDisplayNameSelector)
   const rewardsSenders = useSelector(rewardsSendersSelector)
@@ -123,6 +123,7 @@ export function useTransferFeedDetails(transfer: FeedTokenTransfer) {
   const tokenInfo = useTokenInfoByAddress(transfer.amount.tokenAddress)
   const coinbasePaySenders = useSelector(coinbasePaySendersSelector)
   const fcTransferDisplayInfo = useFiatConnectTransferDisplayInfo(transfer)
+  const isJumpstart = isJumpstartTransaction(transfer)
 
   const {
     type,
@@ -143,6 +144,9 @@ export function useTransferFeedDetails(transfer: FeedTokenTransfer) {
     case TokenTransactionTypeV2.Sent: {
       if (fcTransferDisplayInfo) {
         ;({ title, subtitle, localAmount: customLocalAmount } = fcTransferDisplayInfo)
+      } else if (isJumpstart) {
+        title = t('feedItemJumpstartTitle')
+        subtitle = t('feedItemJumpstartSentSubtitle')
       } else {
         title = t('feedItemSentTitle', { displayName })
         subtitle = t('feedItemSentInfo', { context: !comment ? 'noComment' : null, comment })
@@ -176,6 +180,9 @@ export function useTransferFeedDetails(transfer: FeedTokenTransfer) {
       } else if (isCoinbasePaySender) {
         title = t('feedItemDepositTitle')
         subtitle = t('feedItemReceivedInfo', { context: !comment ? 'noComment' : null, comment })
+      } else if (isJumpstart) {
+        title = t('feedItemJumpstartTitle')
+        subtitle = t('feedItemJumpstartReceivedSubtitle')
       } else {
         title = t('feedItemReceivedTitle', { displayName })
         subtitle = t('feedItemReceivedInfo', { context: !comment ? 'noComment' : null, comment })
